pyproject.toml
[project]
name = "memory-bear"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "weaviate-client>=4.15.4",
    "certifi>=2023.0.0",
    "mcp[cli]>=1.10.1",
    "watchdog>=4.0.0",
    "psutil>=7.0.0",
]
